Move tasks to a separate file
